pyproject.toml
[tool.poetry]
name = "dlmi"
version = "0.1.0"
description = "a university project for deep learning for medical imaging"
authors = ["jonas.milkovits <jonas.milkovits@stud.tu-darmstadt.de>", "frederick wichert <frederick.wichert@stud.tu-darmstadt.de>", "Marlon May <marlon.may@stud.tu-darmstadt.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.2"
torch = "^2.5.1"
torchvision = "^0.20.1"
pyav = "^13.1.0"
lightning = "^2.4.0"
pillow = "^11.0.0"
matplotlib = "^3.10.0"
patchify = "^0.2.3"
scikit-image = "^0.25.0"
tensorboard = "^2.18.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.10.0"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
